+++ common/wrappers/memory-utils.h
